move logic out of capa/render/__init__.py
