+++ scripts/capa_as_library.py
@@ -6,9 +6,10 @@ import collections
 import capa.main
 import capa.rules
 import capa.engine
-import capa.render
 import capa.features
+import capa.render.json
 import capa.render.utils as rutils
+import capa.render.default
 from capa.engine import *
 from capa.render import convert_capabilities_to_result_document
 
@@ -207,9 +208,9 @@ def capa_details(file_path, output_format="dictionary"):
     elif output_format == "json":
         # render results
         # ...as json
-        capa_output = json.loads(capa.render.render_json(meta, rules, capabilities))
+        capa_output = json.loads(capa.render.json.render(meta, rules, capabilities))
     elif output_format == "texttable":
         # ...as human readable text table
-        capa_output = capa.render.render_default(meta, rules, capabilities)
+        capa_output = capa.render.default.render(meta, rules, capabilities)
 
     return capa_output
